esp32 clickEncoder
