+++ platformio.ini
@@ -15,6 +15,7 @@
 [env:lolin32]
 platform = espressif32
 board = lolin32
+lib_ignore=TimerOne
 framework=arduino
-build_flags = -Wno-comment -Wno-reorder -Wno-strict-aliasing -Wno-builtin-macro-redefined -Wno-switch
+build_flags = -Wno-comment -Wno-reorder -Wno-strict-aliasing -Wno-builtin-macro-redefined -Wno-switch -Wno-write-strings
 build_unflags = -Werror=reorder
